Little changes.

+++ manager/filehandler.js
@@ -1,34 +1,78 @@
 const config = require('../config');
 const fs = require('fs');
 const path = require('path');
+const mkdirp = require('mkdirp');
 
 class FileHandler {
   constructor() {
     const dataDir = config.storage.dir.data;
     this.dirs = {
       mp4_1080p: path.join(dataDir, 'mp4_1080p'),
-      jpg_1080p: path.join(dataDir, 'jpg_1080p')
+      webm_1080p: path.join(dataDir, 'webm_1080p'),
+
+      jpg_1080p: path.join(dataDir, 'jpg_1080p'),
+      jpg_1080p_tiny: path.join(dataDir, 'jpg_1080p_tiny'),
+      gif_360p: path.join(dataDir, 'gif_360p')
+    };
+
+    if (!fs.existsSync(this.dirs.webm_1080p)) {
+      mkdirp.sync(this.dirs.webm_1080p);
+    }
+    if (!fs.existsSync(this.dirs.mp4_1080p)) {
+      mkdirp.sync(this.dirs.mp4_1080p);
     }
 
-    if (!fs.existsSync(this.dirs.mp4_1080p)) {
-      fs.mkdirSync(this.dirs.mp4_1080p);
-    }
     if (!fs.existsSync(this.dirs.jpg_1080p)) {
-      fs.mkdirSync(this.dirs.jpg_1080p);
+      mkdirp.sync(this.dirs.jpg_1080p);
+    }
+    if (!fs.existsSync(this.dirs.jpg_1080p_tiny)) {
+      mkdirp.sync(this.dirs.jpg_1080p_tiny);
+    }
+    if (!fs.existsSync(this.dirs.gif_360p)) {
+      mkdirp.sync(this.dirs.gif_360p);
     }
   }
 
-  saveFile(entity, files) {
+  saveFile(entity, files, callback) {
     try {
-      if (files.mp4_1080p) {
+      if (files.mp4_1080p && fs.existsSync(files.mp4_1080p)) {
         fs.renameSync(files.mp4_1080p, path.join(this.dirs.mp4_1080p, entity._id + '.mp4'));
       }
-      if (files.jpg_1080p) {
-        fs.renameSync(files.jpg_1080p, path.join(this.dirs.jpg_1080p, entity._id + '.jpg'));
-      }
-    }catch(err) {
+    } catch(err) {
       console.error(err);
     }
+    try {
+      if (files.webm_1080p && fs.existsSync(files.webm_1080p)) {
+        fs.renameSync(files.webm_1080p, path.join(this.dirs.webm_1080p, entity._id + '.webm'));
+
+      }
+    } catch(err) {
+      console.error(err);
+    }
+
+    try {
+      if (files.jpg_1080p && fs.existsSync(files.jpg_1080p)) {
+        fs.renameSync(files.jpg_1080p, path.join(this.dirs.jpg_1080p, entity._id + '.jpg'));
+      }
+    } catch(err) {
+      console.error(err);
+    }
+    try {
+      if (files.jpg_1080p_tiny && fs.existsSync(files.jpg_1080p_tiny)) {
+        fs.renameSync(files.jpg_1080p_tiny, path.join(this.dirs.jpg_1080p_tiny, entity._id + '.jpg'));
+      }
+    } catch(err) {
+      console.error(err);
+    }
+    try {
+      if (files.gif_360p && fs.existsSync(files.gif_360p)) {
+        fs.renameSync(files.gif_360p, path.join(this.dirs.gif_360p, entity._id + '.gif'));
+      }
+    } catch(err) {
+      console.error(err);
+    }
+
+    callback();
   }
 }
